requirements.txt
asgiref==3.2.3
autopep8==1.4.4
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
cryptography==2.9.2
decorator==4.4.2
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.1.2
django-admin-honeypot==1.1.0
django-allauth==0.41.0
django-crispy-forms==1.7.2
django-filter==2.2.0
django-filters==0.2.1
django-heroku==0.3.1
django-sendgrid-v5==0.8.1
djangorestframework==3.11.0
docx-mailmerge==0.5.0
et-xmlfile==1.0.1
fdfgen==0.16.1
future==0.18.2
gunicorn==20.0.4
httpie==1.0.3
idna==2.8
jdcal==1.4.1
lxml==4.4.2
#mysqlclient==1.4.4
oauthlib==3.1.0
openpyxl==2.6.0
Pillow==6.2.0
psycopg2==2.8.4
psycopg2-binary==2.8.4
pycodestyle==2.5.0
pycparser==2.20
Pygments==2.5.2
PyJWT==1.7.1
PyPDF2==1.26.0
pypng==0.0.20
PyQRCode==1.2.1
python-http-client==1.2.3
python-telegram-bot==12.8
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
sendgrid==2.2.1
sentry-sdk==0.12.3
six==1.12.0
smtpapi==0.3.1
sqlparse==0.3.0
tornado==6.0.4
twilio==6.42.0
urllib3==1.25.7
whitenoise==5.0.1
XlsxWriter==1.2.1
xlwt==1.3.0
